various updates
changed shell script to not recreate files if installation exist

+++ prep.sh
@@ -5,6 +5,7 @@ VARTMP="template.sh"
 VARPMT="ServerStart.sh"
 PROPTMP="server.properties.tmp"
 PROPPMT="server.properties"
+EULA="eula=true"
 URL="https://www.curseforge.com/api/v1/mods/348969/files/2823160/download"
 
 #Function to download modpack
@@ -21,15 +22,15 @@ unzip_modpack() {
 #Moves into data directory
 cd /data
 
-#checks if there is a current modpack installation
-if [ -f "$FILENAME" ]; then
+if [ -e "$FILENAME" ]; then
     echo "Installation Exists"
-    rm -rf $VARPMT
-    rm -rf $PROPPMT
+    #rm -rf $VARPMT
+    #rm -rf $PROPPMT
 
     #Ensures there is a valid variable template file
-    if [ -f "$VARTMP" ]; then
-        cp $VARTMP $VARPMT
+    if [ -e "$VARTMP" ]; then
+        echo "Template Exists"
+        #cp $VARTMP $VARPMT
     else
         echo "Missing $VARTMP file: Unzipping modpack for replacement"
         unzip_modpack
@@ -37,25 +38,33 @@ if [ -f "$FILENAME" ]; then
     fi
 
     #Ensures there is a valid server.properties template file
-    if [ -f "$PROPTMP" ]; then
-        cp $PROPTMP $PROPPMT
+    if [ -e "$PROPTMP" ]; then
+        echo "Properties Exists"
+        #cp $PROPTMP $PROPPMT
     else
         echo "Missing $PROPTMP file: Unzipping modpack for replacement"
         unzip_modpack
         cp $PROPPMT $PROPTMP
     fi
-else
-    
-    #If there is no installation, it downloads the modpack and creates the installation
-    dl_modpack
-    unzip_modpack
-    cp $VARPMT $VARTMP
-    cp $PROPPMT $PROPTMP
 fi
 
 #creates eula.txt file to accept minecraft's EULA
-rm -rf eula.txt
-echo eula=true > eula.txt
+
+#while IFS= read -r line; do
+#    echo "$line"
+#done < eula.txt
+
+read -r line<eula.txt
+
+#echo "$line"
+
+if [[ "$line" == "$EULA" ]]; then
+    echo "Eula True Exists"
+    else
+        echo "Creating Eula"
+        rm -rf eula.txt
+        echo eula=true > eula.txt
+fi
 
 #Sets Variables
 echo ""
